refactor: middleware and remove handlers folder
