+++ src/middleware.ts
@@ -1,10 +1,10 @@
 import { NextResponse } from 'next/server';
 import type { NextRequest } from 'next/server';
 
-import { handleFetchSubscription } from './handlers/subscription';
-import { updateSession } from './libs/supabase/middleware';
-import { createClient } from './libs/supabase/server';
-import SupabaseService from './services/supabase';
+import { updateSession } from '@/libs/supabase/middleware';
+import { createClient } from '@/libs/supabase/server';
+import { fetchSubscription } from '@/services/api/subscription';
+import SupabaseService from '@/services/supabase';
 
 export async function middleware(request: NextRequest) {
   await updateSession(request);
@@ -21,7 +21,8 @@ export async function middleware(request: NextRequest) {
       return NextResponse.redirect(redirectUrl);
     }
 
-    const subscription = await handleFetchSubscription(userId);
+    const subscription = await fetchSubscription(userId);
+
     const plan = subscription &&
       subscription?.status === 'active'
       ? subscription.plan as 'starter' | 'creator' | 'pro'
